// fiftyone/miris-viewer/js/src/looker-3d/types.ts
export type OverlayLabel = { _id: string; path: string; selected: boolean; color?: string; label?: string; sampleId?: string; _cls: string; instance?: { _cls: "Instance"; _id: string }; } & Record<string, unknown>;
// ...
export interface BaseOverlayProps {
  opacity: number;
  rotation: THREE.Vector3Tuple;
  selected: boolean;
  onClick: (e: any) => void;
  label: OverlayLabel;
  color: string;
}

export interface TransformProps extends TransformControlsProps {
  isSelectedForTransform?: boolean;
  onTransformStart?: () => void;
  onTransformEnd?: () => void;
  onTransformChange?: () => void;
  transformControlsRef?: RefObject<any>;
}

export interface HoverState {
  isHovered: boolean;
  setIsHovered: (hovered: boolean) => void;
}

export interface EventHandlers {
  onPointerOver: () => void;
  onPointerOut: () => void;
  onPointerMissed: () => void;
  onPointerMove: (e: ThreeEvent<PointerEvent>) => void;
}